split canonical acceptance specifications from other examples
This explicitely introduces the notion of 'canonical acceptance' (it
has been used in a comment by Joachim, and I think the name is very
clear), and moves most of the examples that had no acc-name into a
subsecent section, for clarity.  The only unamed examples that I have
not moved are the two other classical Rabin encodings, because I feel
that would be less clear out of context.

Note that the Streett-like example I moved in the other acceptance
conditions section is related to some side discussion in #23.  Joachim
and Jan K. both seems to be willing to have a way to name similar
cases.